refactor(core): isolate generators & resolvers
